move json generation code to separate class
